LIBFCREPO-93. Created alto module.
Created an alto.rb module with OCRResource, TextBlock, and TextLine classes to
model the ALTO XML data. Updated the articles.rb script to use the alto module.
Created a lines.rb script that creates an annotation list of all of the text
lines on a given page. Both articles.rb and lines.rb include the text of the
blocks or lines as the text content of their annotations.

https://issues.umd.edu/browse/LIBFCREPO-93